requirements.txt
flask
google-cloud-aiplatform
gunicorn
google-cloud-aiplatform>=1.46.0